Add statistics to pivot cache field.
Field statistics are stored in the file and are available, even if field has no records (save source data = false). Ensure the values are propertly loaded/updated with record values and saved back.

The statistics are not very useful for ClosedXML, but Excel likely uses them for optimized loadings and field structure.