Add and load pivot table:location attributes.
Location attributes contain dimensions that should be loaded to behave as
"dump pipe" and thus should be loaded to fully represent the table.

Load the attributes for location into XLPivotTable. The XLPivotTable.Area
property is kind of redundant wiht Target and not kept in sync. That will
be dealt in further refactoring.