Modify pivot table data structure to be xlsx-like
Original data structure not aligned with xlsx-like and it was necessary to do
rather complicated transaltion from structure in the file to the structure in
ClosedXML during load and again from structure in ClosedXML to xlsx-like
structure during saving.

Because we don't control input files, we have to accept basically anything on
the input which means that transaltion logic would be incredible complex.

Instead, the data structures are modified to allow streamlined loading of files
and we can enable simple pass-through load-save cycle.

This change means, that basically all implemented API had to be reimplemented,
along with loader and saver. The pivot tables API is basically undocumented,
but I tried my best to keep it working (will be deleted soon).

From practical point, we control API and can reject invalid operation and we
can limit possible inputs. That is not possible with files. We have to load
any valid file.

Because of PR size, the pivot styles API wasn't reimplemented, will be done in
a sepearte PR. `XLPivotTable.AllStyleFormats` needs to be fixed and following
tests have been disabled (thus needs to be enabled):

* LoadingTests.CanLoadPivotTableWithBorder
* XLPivotTableTests.PivotTableStyleFormatsTest