Mark IF function as a range function
The IF function should have a scalar test parameter and then two range
parameters. It currently has all parametersmarked as scalar and therefore
reference values are intersected into single values. That is problem when
values from IF function are then used as arguments of some range function
(e.g. SUMIFS or SUM).

Mark the function appropriately, so only the condition parameter is
implicitely intersected.

The test values in test case use logical values directly, because there
is a problem with how implicit intersection is implemented (it should
be done earlier and shoudl depend on the context of the function
argument, but is global instead). This way, the test actually does it
should without running afoul of this problem.
Most users use scalar conditions (e.g. `IF(D1 = "Cake", ..)`), so it is
not a catastrophic problem for now.

Resolves #1707