Add abstration for pivot cache source
There are going to be 4 different sources with wildly different
properties and behavior. It would not be greate to use a single class
for it, as well as stuck all logic and data there.

There is currently only one, so I can get away with casting in the save
method.

Thr equality is required, because that is the only way to deal with
finding a matching cache source when user tries to find matching cache
without exposing various props of individual source types.