Add XLSheetPoint.TryParse
The method will be used to check if ST_CellRef is parseable. Some producers create invalid workbooks and somethimes the behavior should not be "throw an exception" (e.g. <c> can have invalid ST_CelRef and it is still loaded as last cell of row).

It will be used to check that sheet pivot source has a correct area.